done meet & zoom

+++ backend/controllers/zoomHandler.js
@@ -1,6 +1,7 @@
 const jwt = require('jsonwebtoken');
 const axios = require("axios");
 const { promisify } = require("util");
+const db = require("../dbConfig/dbCon");
 // if no token default empty string
 const getTokenOrGenerate = async (prevToken) => {
     if (!prevToken) {
@@ -46,14 +47,61 @@ const checkAndAssignToken = async (req, res) => {
 
 module.exports.create = async (req, res) => {
     try {
+        const { date, time, title, loc, desc, } = req.body;
+        if (!date || !time || !title) {
+            res.status(400).json({
+                status: "failed",
+                message: "date, time, title are required"
+            });
+            return;
+        }
+        const dateReg = new RegExp('([1-9]{1,1}[0-9]{3,3})-([0-1]{1,1}[0-9]{1,1})-([0-1]{1,1}[0-9]{1,1})');
+        const timeReg = new RegExp('([0-9]{1,1}[0-9]{1,1}):([0-9]{1,1}[0-9]{1,1})');
+        if (!dateReg.test(date) || !timeReg.test(time)) {
+            res.status(400).json({
+                status: "failed",
+                message: "improper date or time formats"
+            });
+            return;
+        }
+        let start_time = new Date(date);
+        const [hrs, mins] = time.split(":");
+        start_time.setHours(hrs, mins);
+        let end_time = new Date(start_time);
+        end_time.setMinutes(45);
+        end_time = end_time.getTime();
+        const now = Date.now();
+        const prevMeetRefs = await db.collection('meetups').where('type', '==', 'zmeet').where('start_time', '<=', start_time.getTime()).get();
+        if(!prevMeetRefs.empty){
+            const expDocs = prevMeetRefs.docs;
+            let leave = false;
+            for(let doc of expDocs){
+                const d = doc.data();
+                if(d.end_time <= now){
+                    await db.collection('meetups').doc(doc.id).delete();
+                }
+                if(start_time.getTime() < d.end_time){
+                    leave = true;
+                }
+            }
+            if(leave){
+            res.status(400).json({
+                status:"failed",
+                message:"Busy time"
+            });
+            return;
+        }
+        }
         const token = await checkAndAssignToken(req, res);
         const URL = `https://api.zoom.us/v2/users/me/meetings`;
         const data = JSON.parse(JSON.stringify({
-            "topic": req.body.topic,
-            "start_time": req.body.start_time,
-            "password": req.body.password,
-            "agenda": req.body.agenda
+            "topic": title,
+            "start_time": start_time.toLocaleString(),
+            duration: 45,
+            timezone: 'Asia/Kolkata',
+            "agenda": desc
         }));
+        start_time = start_time.getTime();
         const zoomRes = await axios.post(URL, data, {
             "headers": {
                 'User-Agent': 'Zoom-api-Jwt-Request',
@@ -61,9 +109,26 @@ module.exports.create = async (req, res) => {
                 "Authorization": `Bearer ${token}`
             }
         });
+        const meetData = {
+            type: 'zmeet',
+            eventId: zoomRes.data.id,
+            meetLink: zoomRes.data.start_url,
+            cred_req: true,
+            password: zoomRes.data.password,
+            title: zoomRes.data.topic,
+            desc: zoomRes.data.agenda,
+            date: date,
+            time: time,
+            start_time: start_time,
+            end_time: end_time,
+            duration: zoomRes.data.duration + "min",
+            created_on: Date.now()
+        }
+        const meetDocRef = await db.collection('meetups').add(JSON.parse(JSON.stringify(meetData)));
         res.status(parseInt(zoomRes.status)).json({
-            status: zoomRes.statusText,
-            data: zoomRes.data
+            status: "success",
+            id: meetDocRef.id,
+            data: meetData
         });
     } catch (error) {
         console.log(error)
@@ -106,15 +171,25 @@ module.exports.getAll = async (req, res) => {
 
 module.exports.delete = async (req, res) => {
     try {
-        const mid = req.params.mid;
+        const mid = req.params.id;
+        if (!mid) {
+            res.status(400).json({
+                status: "failed",
+                message: "provide id"
+            })
+            return;
+        }
+        const meetRef = await db.collection("meetups").doc(mid).get();
+        if (!meetRef.exists) {
+            res.status(400).json({
+                status: "failed",
+                message: "No such meet exists"
+            })
+            return;
+        }
+        const { eventId } = meetRef.data();
         const token = await checkAndAssignToken(req, res);
-        const URL = "https://api.zoom.us/v2/meetings/" + mid;
-        const data = JSON.parse(JSON.stringify({
-            "topic": req.body.topic,
-            "start_time": req.body.start_time,
-            "password": req.body.password,
-            "agenda": req.body.agenda
-        }));
+        const URL = "https://api.zoom.us/v2/meetings/" + eventId;
         const zoomRes = await axios.delete(URL, {
             "headers": {
                 'User-Agent': 'Zoom-api-Jwt-Request',
@@ -122,10 +197,10 @@ module.exports.delete = async (req, res) => {
                 'Authorization': `Bearer ${token}`
             }
         });
-        console.log(res.body);
-        res.status(parseInt(zoomRes.status)).json({
+        const delMeet = await db.collection("meetups").doc(mid).delete();
+        res.status(200).json({
             status: "success",
-            data: "Deleted zoom meeting "+mid
+            data: "Deleted zoom meeting " + mid
         });
     } catch (error) {
         console.log(error)
